+++ main.py
@@ -210,6 +210,9 @@ class ReverseSearcherPlugin(Star):
             proxies=config.get("proxies", ""),
             timeout=60,
             default_params=default_params,
+            allow_third_party_image_host=config.get(
+                "allow_third_party_image_host", True
+            ),
         )
         self.state_handlers = {
             "waiting_engine": self._handle_waiting_engine,